[#6]Design: 컴포넌트 분리
