+++ src/components/Main/Section/Section.styles.ts
@@ -1,0 +1,29 @@
+import styled from '@emotion/styled';
+
+export const Head = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 8px 16px;
+  margin-top: 8px;
+`;
+
+export const Title = styled.h2`
+  margin: 0;
+  font-size: 16px;
+  font-weight: 800;
+`;
+
+export const Add = styled.button`
+  border: 0;
+  background: transparent;
+  font-size: 20px;
+  color: black;
+`;
+
+export const List = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  padding: 8px 16px 16px;
+`;
