+++ src/pages/Main/MainPage.tsx
@@ -3,8 +3,11 @@ import * as S from './MainPage.styles';
 import SearchBar from '../../components/Main/SearchBar/SearchBar';
 import Banner from '../../components/Main/Banner/Banner';
 import QuickActions from '../../components/Main/QuickActions/QuickActions';
-import { SectionHeader, SectionList } from '../../components/Any/Section';
-import NoticeCard from '../../components/Any/NoticeCard';
+import {
+  SectionHeader,
+  SectionList,
+} from '../../components/Main/Section/Section';
+import NoticeCard from '../../components/Card/NoticeCard';
 import { latest, dueSoon } from '../../data/notices';
 import FabChat from '../../components/FabChat';
 
@@ -17,7 +20,7 @@ export default function MainPage() {
           <SearchBar />
           <Banner />
           <QuickActions />
-          
+
           <SectionHeader title="최신 공고" />
           <SectionList>
             {latest.map((n) => (
